Probably command ui can be useful in other apps. Split it!
